backend/requirements/base.txt
Django==5.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.2.1
python-dotenv==1.1.0
